Create separate cross chain governance scenario file
